- Drop Chance is a double now so that decimals are supported

+++ LootTable.cs
@@ -27,14 +27,14 @@ namespace LootControl
   {
     public int ItemID;
     public int PrefixID;
-    public int DropChance = 100;
+    public double DropChance = 100;
     public int Min_Stack = 0;
     public int Max_Stack = 1;
     public bool DropInClassic = true;
     public bool DropInExpert = false;
     public bool DropInMaster = false;
 
-    public Loot(int pItemID, int pDropChance, int pMin_Stack, int pMax_Stack, int pPrefixID=0, bool pDropInClassic=true, bool pDropInExpert=false, bool pDropInMaster=false)
+    public Loot(int pItemID, double pDropChance, int pMin_Stack, int pMax_Stack, int pPrefixID=0, bool pDropInClassic=true, bool pDropInExpert=false, bool pDropInMaster=false)
     {
       ItemID = pItemID;
       PrefixID = pPrefixID;
